[procon test] Run test src/atcoder/ddcc2019-qual/d.cc

+++ src/atcoder/ddcc2019-qual/d.cc
@@ -943,7 +943,7 @@ static pair<Vector<i64>, Vector<bool>> sieve(i64 n) {
 
 void body() {
     auto as = read<i64>(30);
-    Vector<int> x(30);
+    Vector<i64> x(30);
     REP (i, 30) {
         x[i] = i + 1;
     }
